Refactored Persistence Namespace to Data.Interfaces and .Data.Repositories and added UnitOfWork prep work. This also touched the JokesAPI.Tests Project.
